feature (comments) comment to a post
checking if a post exist
check if a post is published
add a comment

[Delivers #163520287]
