+++ server/src/routes/index.js
@@ -1,6 +1,5 @@
 import auth from './auth';
 import users from './users';
 import posts from './posts';
-import comments from './comments';
 
-export { auth, users, posts, comments };
+export { auth, users, posts };
